Create Lanternfish_day6-part2.cpp
credits to Partman for this awesome method.

instead of simulating the array, the obvious bruteforce method, which works for n~160 at max taking ~25s to execute, we only actually need the number of fishes at a particular day. for this approach, i have indexed the number of fishes at a given time, according to their internal timer. so 3,4,3,1,2 can be better represented as 011210000, the next day 2,3,2,0,1 can be obtained by simply shifting the elements leftward once. this obviously needs to compensate by adding extra 6s for every 0s lost. in this method 8s are managed automatically since the 0s are being rotated towards 8s. hence, this runs in O(n) complexity where n=number of days, instead of O(2^n). the answer would be simply the sum of the index array. this method, for comparison, takes ~1s. another thing to be careful of is integer overflows. long long int and BigInteger should help with this